progress on install-nextcloud.sh

+++ install-nextcloud.sh
@@ -13,29 +13,39 @@ done
 
 # installs prerequisites
 
-sudo apt update
-sudo apt -y install apache2 mariadb-server libapache2-mod-php7.4
-sudo apt -y install php7.4-gd php7.4-mysql php7.4-curl php7.4-mbstring php7.4-intl
-sudo apt -y install php7.4-gmp php7.4-bcmath php-imagick php7.4-xml php7.4-zip
+FLAGS_DIR=/home/$USER/.flags
+if [ ! -d $FLAGS_DIR ]; then
+	mkdir $FLAGS_DIR
+fi
 
-echo
-echo
-echo "Starting MySQL command line mode."
-echo "Press the enter key when prompted for a password."
-echo "Once the prompt appears, enter the following commands,"
-echo "replacing username and password with appropriate values:"
-echo
-echo CREATE USER 'username'@'localhost' IDENTIFIED BY 'password';
-echo CREATE DATABASE IF NOT EXISTS nextcloud CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;
-echo GRANT ALL PRIVILEGES ON nextcloud.* TO 'username'@'localhost';
-echo FLUSH PRIVILEGES;
-echo quit;
-echo
-sudo mysql -uroot -p
+if [ ! -e $FLAGS_DIR/install-nextcloud-prerequisites ]; then
+
+	sudo apt update
+	sudo apt -y install apache2 mariadb-server libapache2-mod-php7.4
+	sudo apt -y install php7.4-gd php7.4-mysql php7.4-curl php7.4-mbstring php7.4-intl
+	sudo apt -y install php7.4-gmp php7.4-bcmath php-imagick php7.4-xml php7.4-zip
+
+	echo
+	echo
+	echo "Starting MySQL command line mode."
+	echo "Press the enter key when prompted for a password."
+	echo "Once the prompt appears, enter the following commands,"
+	echo "replacing username and password with appropriate values:"
+	echo
+	echo CREATE USER 'username'@'localhost' IDENTIFIED BY 'password';
+	echo CREATE DATABASE IF NOT EXISTS nextcloud CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;
+	echo GRANT ALL PRIVILEGES ON nextcloud.* TO 'username'@'localhost';
+	echo FLUSH PRIVILEGES;
+	echo quit;
+	echo
+	sudo mysql -uroot -p
+
+	touch $FLAGS_DIR/install-nextcloud-prerequisites
+fi
 
 
-echo "pausing here:"
-
+echo "pausing here"
+exit 0
 
 # downloads NextCloud
